+++ project_addons/theme_rias_menu/__manifest__.py
@@ -3,15 +3,19 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 {
     'name': 'Theme Rías Menu',
-    'version': '11.0.0.0.0',
+    'version': '11.0.1.0.0',
+    'summary': 'Add an Snippet Content Menu by customizable view',
     'category': 'Custom',
     'license': 'AGPL-3',
     'author': "Comunitea, ",
     'depends': [
+        # Custom
         'theme_rias',
+        # Odoo
+        'website',
     ],
     'data': [
-        'data/menu_data.xml',
+        # 'data/menu_data.xml',
         'views/website_menu_view.xml',
         'templates/snippets.xml',
         'templates/header.xml',