+++ project_addons/theme_rias/doc/changelog.rst
@@ -1,3 +1,9 @@
+`11.0.1.0.0`
+------------
+- **[UDP]** Menu data from Theme Rias Menu (Keep menu_icon field to show menus)
+- **[ADD]** menu_icon field for menu model to show menus on Theme Rias Menu
+- **[DEL]** Templates, views and snippets for Snippet Content Menu
+
 `11.0.0.0.2`
 ------------
 - **[UDP]** Make logo work with website_logo module by ecommerce_base module